requirements.txt
appdirs==1.4.4
asgiref==3.4.1
baron==0.9
black==21.7b0
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.4
click==8.0.1
colorama==0.4.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.4.7
defusedxml==0.7.1
dj-database-url==0.5.0
dj-email-url==1.0.2
dj-rest-auth==1.1.0
Django==3.2.6
django-allauth==0.45.0
django-cache-url==3.2.3
django-cors-headers==3.4.0
django-environ==0.4.5
django-environs==0.1.1
django-heroku==0.3.1
django-js-asset==1.2.2
django-mptt==0.12.0
djangorestframework==3.12.4
drf-yasg==1.20.0
environs==9.3.2
gunicorn==20.1.0
idna==3.2
inflection==0.5.1
install==1.3.4
isort==5.9.3
itypes==1.2.0
Jinja2==3.0.1
MarkupSafe==2.0.1
marshmallow==3.13.0
mypy-extensions==0.4.3
oauthlib==3.1.1
packaging==21.0
pathspec==0.9.0
Pillow==8.3.1
psycopg2==2.9.1
psycopg2-binary==2.9.1
pycparser==2.20
PyJWT==2.1.0
pyparsing==2.4.7
python-dotenv==0.19.0
python3-openid==3.2.0
pytz==2021.1
PyYAML==5.3.1
redbaron==0.9.2
regex==2021.7.6
requests==2.26.0
requests-oauthlib==1.3.0
rply==0.7.8
ruamel.yaml==0.17.10
ruamel.yaml.clib==0.2.6
six==1.16.0
sqlparse==0.4.1
tomli==1.2.0
uritemplate==3.0.1
urllib3==1.26.6
whitenoise==5.3.0
